Add Honeybadger::Dependency
In order to start moving our 3rd party integration gems into honeybadger-ruby, we need a way to detect if the dependent libraries are present in a project installation. This adds Honeybadger::Dependency, which provides the following API for registering dependencies:

module Honeybadger
  Dependency.register do
    requirement { defined?(Delayed::Plugins::Plugin) }
    requirement { defined?(Delayed::Worker.plugins) }

    injection do
      require 'honeybadger/integrations/delayed_job/plugin'
      Delayed::Worker.plugins << Integrations::DelayedJob::Plugin
    end
  end
end

The code in the injection block is executed only when all requirement blocks return true. For the code to be injected, Honeybadger::Dependency.inject! must be called *after* all 3rd-party libs have been required.